move flake8 test
