+++ internal/carpentry/carpentry_test.go
@@ -4,14 +4,10 @@ import (
 	"bytes"
 	"github.com/arcalot/arcaflow-plugin-image-builder/internal/carpentry"
 	"github.com/arcalot/arcaflow-plugin-image-builder/internal/dto"
-	"github.com/arcalot/arcaflow-plugin-image-builder/internal/requirements"
 	mock_ces "github.com/arcalot/arcaflow-plugin-image-builder/mocks/ce_service"
 	"github.com/golang/mock/gomock"
 	"go.arcalot.io/assert"
 	arcalog "go.arcalot.io/log"
-	"log"
-	"os"
-	"path/filepath"
 	"testing"
 )
 
@@ -67,27 +63,4 @@ func TestAllTrue(t *testing.T) {
 func TestCliCarpentry(t *testing.T) {
 	logger := arcalog.NewLogger(arcalog.LevelInfo, arcalog.NewNOOPLogger())
 	assert.Error(t, carpentry.CliCarpentry(true, true, logger, "podman"))
-}
-
-func TestFlake8(t *testing.T) {
-	stdout := &bytes.Buffer{}
-	logger := arcalog.New(arcalog.Config{
-		Level:       arcalog.LevelInfo,
-		Destination: arcalog.DestinationStdout,
-		Stdout:      os.Stdout,
-	})
-	err := requirements.Flake8PythonCodeStyle("/githug/workplace", stdout, logger)
-	assert.Error(t, err)
-
-	afp, patherr := filepath.Abs("../../fixtures/pep8_compliant")
-	if patherr != nil {
-		log.Fatal(patherr)
-	}
-	assert.Nil(t, requirements.Flake8PythonCodeStyle(afp, &bytes.Buffer{}, logger))
-
-	afp, patherr = filepath.Abs("../../fixtures/pep8_non_compliant")
-	if patherr != nil {
-		log.Fatal(patherr)
-	}
-	assert.Error(t, requirements.Flake8PythonCodeStyle(afp, &bytes.Buffer{}, logger))
 }
